+++ DragonShield/DatabaseManager.swift
@@ -39,6 +39,8 @@ class DatabaseManager: ObservableObject {
     @Published var dbFilePath: String = ""
     @Published var dbCreated: Date?
     @Published var dbModified: Date?
+    @Published var productionDBPath: String = ""
+    @Published var testDBPath: String = ""
     // Add other config items as @Published if they need to be globally observable
     // For fx_api_provider, fx_update_frequency, we might just display them or use TextFields